Refactor into assets folder
