+++ packages/components/Input/src/assets/dismissSvg.tsx
