+++ packages/components/Input/src/index.ts
@@ -1,3 +1,4 @@
 export { Input, inputLookup } from './Input';
 export type { InputProps, InputSlotProps, InputTokens, InputType } from './Input.types';
 export { input } from './Input.types';
+export { DismissSvg } from './assets/dismissSvg';
